Move internal stuff and `CallbackResult` class to the core module
