+++ aioresponses/core.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 import asyncio
+import json
 from collections import namedtuple
+from distutils.version import StrictVersion
 from functools import wraps
 from typing import Callable, Dict, Tuple, Union, Optional, List  # noqa
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 from aiohttp import (
     ClientConnectionError,
@@ -12,14 +14,37 @@ from aiohttp import (
     hdrs,
     http
 )
+from aiohttp.helpers import TimerNoop
+from multidict import CIMultiDict
 
 from .compat import (
+    AIOHTTP_VERSION,
     URL,
     Pattern,
+    stream_reader_factory,
     merge_params,
     normalize_url,
-    build_response,
 )
+
+
+class CallbackResult:
+
+    def __init__(self, method: str = hdrs.METH_GET,
+                 status: int = 200,
+                 body: str = '',
+                 content_type: str = 'application/json',
+                 payload: Dict = None,
+                 headers: Dict = None,
+                 response_class: 'ClientResponse' = None,
+                 reason: Optional[str] = None):
+        self.method = method
+        self.status = status
+        self.body = body
+        self.content_type = content_type
+        self.payload = payload
+        self.headers = headers
+        self.response_class = response_class
+        self.reason = reason
 
 
 class RequestMatch(object):
@@ -74,31 +99,86 @@ class RequestMatch(object):
             return False
         return self.match_func(url)
 
-    async def _build_response(
+    def _build_raw_headers(self, headers: Dict) -> Tuple:
+        """
+        Convert a dict of headers to a tuple of tuples
+
+        Mimics the format of ClientResponse.
+        """
+        raw_headers = []
+        for k, v in headers.items():
+            raw_headers.append((k.encode('utf8'), v.encode('utf8')))
+        return tuple(raw_headers)
+
+    def _build_response(self, url: 'Union[URL, str]',
+                        method: str = hdrs.METH_GET,
+                        status: int = 200,
+                        body: str = '',
+                        content_type: str = 'application/json',
+                        payload: Dict = None,
+                        headers: Dict = None,
+                        response_class: 'ClientResponse' = None,
+                        reason: Optional[str] = None) -> ClientResponse:
+        if response_class is None:
+            response_class = ClientResponse
+        if payload is not None:
+            body = json.dumps(payload)
+        if not isinstance(body, bytes):
+            body = str.encode(body)
+        kwargs = {}
+        if AIOHTTP_VERSION >= StrictVersion('3.1.0'):
+            loop = Mock()
+            loop.get_debug = Mock()
+            loop.get_debug.return_value = True
+            kwargs['request_info'] = Mock()
+            kwargs['writer'] = Mock()
+            kwargs['continue100'] = None
+            kwargs['timer'] = TimerNoop()
+            if AIOHTTP_VERSION < StrictVersion('3.3.0'):
+                kwargs['auto_decompress'] = True
+            kwargs['traces'] = []
+            kwargs['loop'] = loop
+            kwargs['session'] = None
+        # We need to initialize headers manually
+        _headers = CIMultiDict({hdrs.CONTENT_TYPE: content_type})
+        if headers:
+            _headers.update(headers)
+        raw_headers = self._build_raw_headers(_headers)
+        resp = response_class(method, url, **kwargs)
+        if AIOHTTP_VERSION >= StrictVersion('3.3.0'):
+            # Reified attributes
+            resp._headers = _headers
+            resp._raw_headers = raw_headers
+        else:
+            resp.headers = _headers
+            resp.raw_headers = raw_headers
+        resp.status = status
+        resp.reason = reason
+        resp.content = stream_reader_factory()
+        resp.content.feed_data(body)
+        resp.content.feed_eof()
+        return resp
+
+    async def build_response(
             self, url: URL, **kwargs: Dict
     ) -> 'Union[ClientResponse, Exception]':
         if isinstance(self.exception, Exception):
             return self.exception
         if callable(self.callback):
-            resp = self.callback(url, **kwargs)
+            result = self.callback(url, **kwargs)
         else:
-            resp = None
-        if resp is None:
-            url = url
-            resp_data = self
-        else:
-            url = resp.url
-            resp_data = resp
-        resp = build_response(
+            result = None
+        result = self if result is None else result
+        resp = self._build_response(
             url=url,
-            method=resp_data.method,
-            status=resp_data.status,
-            body=resp_data.body,
-            content_type=resp_data.content_type,
-            payload=resp_data.payload,
-            headers=resp_data.headers,
-            response_class=resp_data.response_class,
-            reason=resp_data.reason)
+            method=result.method,
+            status=result.status,
+            body=result.body,
+            content_type=result.content_type,
+            payload=result.payload,
+            headers=result.headers,
+            response_class=result.response_class,
+            reason=result.reason)
         return resp
 
 
@@ -218,7 +298,7 @@ class aioresponses(object):
     ) -> Optional['ClientResponse']:
         for i, matcher in enumerate(self._matches):
             if matcher.match(method, url):
-                response = await matcher._build_response(url, **kwargs)
+                response = await matcher.build_response(url, **kwargs)
                 break
         else:
             return None
